requirements.txt
fastapi
uvicorn
streamlit
openai
groq
langchain-groq
langchain-pinecone
langchain
langchain-openai
twilio
gtts
pydub
python-dotenv
pypdf
faiss-cpu
sentence-transformers
pandas
sentence-transformers
langchain_community
tavily-python
elevenlabs
streamlit-mic-recorder
langdetect